# assistant/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Assistant backend service API",
        "title": "Assistant API",
        "contact": {},
        "version": "1.0"
    },
    "host": "127.0.0.1:4321",
    "basePath": "/",
    "paths": {
        "/api/filebrowser/copy": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "文件浏览"
                ],
                "summary": "复制文件/目录到目标目录",
                "parameters": [
                    {
                        "description": "{paths, dest}",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/filebrowser/delete": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "文件浏览"
                ],
                "summary": "删除文件/目录",
                "parameters": [
                    {
                        "description": "{paths}",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/filebrowser/download": {
            "get": {
                "tags": [
                    "文件浏览"
                ],
                "summary": "下载文件",
                "parameters": [
                    {
                        "type": "string",
                        "description": "相对 root 的文件路径",
                        "name": "path",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    }
                }
            }
        },
        "/api/filebrowser/download-tgz": {
            "get": {
                "description": "将指定路径列表(文件或目录)打包为 tar.gz 下载",
                "tags": [
                    "文件浏览"
                ],
                "summary": "下载选中文件/目录为 tar.gz",
                "parameters": [
                    {
                        "type": "string",
                        "description": "路径(可重复 path=a\u0026path=b)",
                        "name": "paths",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    }
                }
            }
        },
        "/api/filebrowser/list": {
            "get": {
                "tags": [
                    "文件浏览"
                ],
                "summary": "列出目录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "相对 root 的路径(空=root)",
                        "name": "path",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/filebrowser/mkdir": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "文件浏览"
                ],
                "summary": "创建目录",
                "parameters": [
                    {
                        "description": "{path}",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/filebrowser/move": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "文件浏览"
                ],
                "summary": "移动文件/目录到目标目录",
                "parameters": [
                    {
                        "description": "{paths, dest}",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/filebrowser/raw": {
            "get": {
                "tags": [
                    "文件浏览"
                ],
                "summary": "在线预览文件",
                "parameters": [
                    {
                        "type": "string",
                        "description": "相对 root 的文件路径",
                        "name": "path",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "1=生成缩略图(仅图片)",
                        "name": "thumb",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    }
                }
            }
        },
        "/api/filebrowser/rename": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "文件浏览"
                ],
                "summary": "重命名文件/目录",
                "parameters": [
                    {
                        "description": "{path, new_name}",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/filebrowser/touch": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "文件浏览"
                ],
                "summary": "创建空文件（不覆盖）",
                "parameters": [
                    {
                        "description": "{path}",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
        "/api/filebrowser/trash/delete": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "文件浏览"
                ],
                "summary": "从回收站永久删除",
                "parameters": [
                    {
                        "description": "{trash_names}",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/filebrowser/trash/restore": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "文件浏览"
                ],
                "summary": "从回收站恢复",
                "parameters": [
                    {
                        "description": "{trash_names}",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/filebrowser/upload": {
            "post": {
                "description": "multipart/form-data 上传; 字段 file 为文件(可多个), path 为目标目录(相对 root, 空=root)",
                "consumes": [
                    "multipart/form-data"
                ],
                "tags": [
                    "文件浏览"
                ],
                "summary": "上传文件(不覆盖，支持多文件)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "目标目录(相对 root)",
                        "name": "path",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "文件(可多个)",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
    }
}